Add files via upload
I connected both computers to the router. The arrow is facing the router because you would be connecting the computers to the router. I connected both computers to a printer. The arrow is going from the computer to the printer because you would be on your computers operating system connecting to the printer. I put a connection between the printer and the router because the printer must have an internet connection.